requirements.txt
apng==0.3.4
asgiref==3.8.1
bcrypt==4.2.0
binwalk==2.1.0
boto3==1.38.9
botocore==1.38.9
capstone==6.0.0a1
cement==3.0.12
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colored-traceback==0.4.2
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
Cython==3.0.12
decorator==5.1.1
deluge-client==1.10.2
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.6
django-cors-headers==4.6.0
django-jazzmin==3.0.1
django-storages==1.14.6
docopt==0.6.2
docx2pdf==0.1.8
ExifRead==3.0.0
fonttools==4.57.0
git-filter-repo==2.47.0
gmpy2==2.2.1
gunicorn==23.0.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.7.0
intervaltree==3.1.0
jmespath==1.0.1
kiwisolver==1.4.8
lxml==5.3.0
Mako==1.3.6
MarkupSafe==3.0.2
matplotlib==3.10.1
morse3==2.9
moviepy==2.1.1
mpmath==1.3.0
mysqlclient==2.2.4
numpy==2.2.0
packaging==24.1
pandas==2.2.3
paramiko==3.5.0
pillow==10.4.0
plumbum==1.9.0
prettytable==3.14.0
proglog==0.1.10
psutil==6.1.0
psycopg2-binary==2.9.9
pwn==1.0
pwntools==4.14.0
pycparser==2.22
pycryptodome==3.21.0
pyelftools==0.31
Pygments==2.18.0
PyMuPDF==1.24.13
PyNaCl==1.5.0
pyparsing==3.2.3
PyPDF2==3.0.1
pypng==0.20220715.0
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
python-magic==0.4.27
python-xmp-toolkit==2.0.2
pytube==15.0.0
pytz==2025.1
requests==2.32.3
ROPGadget==7.5
rpyc==6.0.1
s3transfer==0.12.0
sagemath==1.3.0
scipy==1.15.2
setuptools==75.6.0
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.5.3
stegoveritas==1.10
stegoveritas-binwalk==2.1.2
stegoveritas-pfp==0.2.3
stegoveritas-py010parser==0.1.10
sympy==1.13.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
unicorn==2.1.1
unix-ar==0.2.1
urllib3==2.2.3
wcwidth==0.2.13
xortool==1.0.2
yt-dlp==2024.11.18
zipp==3.21.0
zstandard==0.23.0
